# Speed up & clean up host layout conversion 