--- a/tt_metal/api/tt-metalium/tilize_utils.hpp
+++ b/tt_metal/api/tt-metalium/tilize_utils.hpp
@@ -19,8 +19,8 @@ namespace tests::utils {
 enum class TensorLayoutType {
     LIN_ROW_MAJOR = 0,   // standard element-wise row-major
     TILED_SWIZZLED = 1,  // row-major of tiles, each tile is row-major-swizzled
-    TILED_NFACES =
-        2,  // row-major of tiles, each tile is N (N = 1, 2, or 4) faces, each face is row-major, faces are swizzled
+    TILED_NFACES = 2,    // row-major of tiles, each tile is N (N = 1, 2, or 4) faces, each face is
+                         // row-major, faces are swizzled
 };
 std::ostream& operator<<(std::ostream& os, TensorLayoutType layout);
 }  // namespace tests::utils