### Ticket
Github issue: https://github.com/tenstorrent/tt-metal/issues/8621 
Previous incomplete PR: https://github.com/tenstorrent/tt-metal/pull/17848 

### Problem description
Host layout conversion is not optimal.

### What's changed
- Rewrite host layout conversion functions with better memory coalescing and preallocation of buffers.
- Add direct `LIN_ROW_MAJOR` <--> `TILED_NFACES` conversion instead of two conversions
- Remove most duplications of tilization/untilization in the code. 
- Add proper C++ test for layout conversions (different shapes, different types, different layouts)
- Add benchmark to properly compare previous and current implementations in terms of speed

#### Benchmarking:
Shape: (10240, 1024)
Biggest difference in Release is LIN_ROW_MAJOR --> TILED_NFACES (78.7ms -> 26.2ms = 3.00x speed up)
FYI: LIN_ROW_MAJOR --> TILED_NFACES == `ttnn.to_layout(row_major_tensor, ttnn.TILE_LAYOUT)`

**Release**
Before PR:
```
./build/tt_metal/benchmark/host_tilizer_untilizer/tilizer_untilizer --benchmark_min_time=10
--------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations
--------------------------------------------------------------------------------------
BM_ConvertLayout_RowMajorToTiledSwizzled          26.3 ms         26.3 ms          490
BM_ConvertLayout_RowMajorToTiledNfaces            78.7 ms         78.7 ms          176
BM_ConvertLayout_TiledSwizzledToRowMajor          25.5 ms         25.4 ms          529
BM_ConvertLayout_TiledSwizzledToTiledNFaces       51.9 ms         51.9 ms          268
BM_ConvertLayout_TiledNFacesToRowMajor            65.7 ms         65.7 ms          217
BM_ConvertLayout_TiledNFacesToTiledSwizzled       39.0 ms         39.0 ms          357
```

After PR: 
```
./build/tt_metal/benchmark/host_tilizer_untilizer/tilizer_untilizer --benchmark_min_time=10
--------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations
--------------------------------------------------------------------------------------
BM_ConvertLayout_RowMajorToTiledSwizzled          25.2 ms         25.2 ms          533
BM_ConvertLayout_RowMajorToTiledNfaces            26.2 ms         26.2 ms          552
BM_ConvertLayout_TiledSwizzledToRowMajor          25.4 ms         25.4 ms          549
BM_ConvertLayout_TiledSwizzledToTiledNFaces       27.3 ms         27.3 ms          535
BM_ConvertLayout_TiledNFacesToRowMajor            25.3 ms         25.3 ms          595
BM_ConvertLayout_TiledNFacesToTiledSwizzled       22.2 ms         22.2 ms          65
```

**Debug** -- less useful than release, but difference is huge, might be noticeable during development

Before PR:
```
./build/tt_metal/benchmark/host_tilizer_untilizer/tilizer_untilizer --benchmark_min_time=10
--------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations
--------------------------------------------------------------------------------------
BM_ConvertLayout_RowMajorToTiledSwizzled           145 ms          145 ms           97
BM_ConvertLayout_RowMajorToTiledNfaces             389 ms          389 ms           37
BM_ConvertLayout_TiledSwizzledToRowMajor           130 ms          130 ms          108
BM_ConvertLayout_TiledSwizzledToTiledNFaces        240 ms          240 ms           58
BM_ConvertLayout_TiledNFacesToRowMajor             278 ms          278 ms           51
BM_ConvertLayout_TiledNFacesToTiledSwizzled        145 ms          145 ms           97
```

After PR:
```
./build/tt_metal/benchmark/host_tilizer_untilizer/tilizer_untilizer --benchmark_min_time=10
--------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations
--------------------------------------------------------------------------------------
BM_ConvertLayout_RowMajorToTiledSwizzled          42.1 ms         42.1 ms          331
BM_ConvertLayout_RowMajorToTiledNfaces            55.2 ms         55.2 ms          252
BM_ConvertLayout_TiledSwizzledToRowMajor          41.3 ms         41.3 ms          397
BM_ConvertLayout_TiledSwizzledToTiledNFaces       64.5 ms         64.5 ms          170
BM_ConvertLayout_TiledNFacesToRowMajor            42.8 ms         42.8 ms          321
BM_ConvertLayout_TiledNFacesToTiledSwizzled       36.4 ms         36.4 ms          385
```

### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes
- [ ] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI passes (if applicable)
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml) CI passes (if applicable)
- [ ] New/Existing tests provide coverage for changes