@philei-tt 2025-04-08
PR is very bulky, since I update all calls to layout conversion. 
The main changes are 
- update layout conversion code: [tt_metal/impl/data_format/tilize_utils.cpp](https://github.com/tenstorrent/tt-metal/pull/20335/files#diff-f9197ee49dd22c652447ee454a6e333558df993a9ce32f0176a3014ffc8692a8), [tt_metal/api/tt-metalium/tilize_utils.hpp](https://github.com/tenstorrent/tt-metal/pull/20335/files#diff-e61fde88c7c3fa5d5e58296f74d7d1d1aa9737965acd9b29954e9eebd40254b1)
- new tests: [tests/tt_metal/tt_metal/api/test_tilize_untilize.cpp](https://github.com/tenstorrent/tt-metal/pull/20335/files#diff-ddfaaae7224f9f1a5047428823bf3fe7ca4b0123ec95edb3293fa21f877fb2da) 
- New benchmark: [tests/tt_metal/tt_metal/benchmark/host_tilizer_untilizer/tilizer_untilizer.cpp](https://github.com/tenstorrent/tt-metal/pull/20335/files#diff-348c46c06d699c38ded2dbf65fe7a7d655ec8925f7947db21ba8ecd2deafd0d8)